added new versions of the css, and html as well as the icon file
I worked on the front side of the project and changed all the files so they could work together in the same directory. Just download and place them all in the same directory and they should work. I just used my dreamweaver editor to do this, no node or other things. You guys can change, add, delete or whatever to this as you see fit. Dave